requirements.txt
flask==2.2.3
pymongo==4.3.3
flask-login==0.6.2
python-dotenv==1.0.0
flask-wtf==1.1.1
bcrypt==4.0.1
bson==0.5.10
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.2
dnspython==2.3.0